+++ inc/config/http-query-context/http-query-context.php
@@ -20,6 +20,6 @@ interface HttpQueryContext {
 	public function get_request_body( array $input_variables ): array|null;
 	public function get_query_name(): string;
 	public function get_query_runner(): QueryRunnerInterface;
-	public function get_results( string $response_data, array $input_variables ): array;
 	public function is_collection(): bool;
+	public function process_response( string $raw_response_data, array $input_variables ): string;
 }
